Add styling to some components.. - Add css modules to components - Remove inline css in some components - Fix error page and add "goback" button
