+++ src/components/FiveDayForecast/FiveDay.module.css
@@ -1,0 +1,28 @@
+.FiveDayContainer {
+    display: flex;
+    width: 45vw;
+    flex-wrap: wrap;
+    flex-direction: column;
+    position: relative;
+    margin: 0 auto;
+    text-transform: capitalize;
+    border: 1px solid gray;
+    border-radius: 10px;
+    padding: 1rem;
+  }
+  
+.row {
+    display: flex;
+    justify-content: space-evenly;
+    align-items: center;
+    border-bottom: 1px solid lightgray;
+  }
+  
+.minTemp {
+    color: blue;
+  }
+  
+.maxTemp {
+    color: red;
+  }
+
