+++ src/components/FiveDayForecast/FiveDayForecast.jsx
@@ -1,50 +1,23 @@
 import React from 'react';
-
-const fiveDayStyle = {
-  display: 'flex',
-  // justifyContent: 'space-between',
-  width: '45vw',
-  flexWrap: 'wrap',
-  flexDirection: 'column',
-  position: 'relative',
-  margin: '0 auto',
-  textTransform: 'capitalize',
-  border: '1px solid gray',
-  borderRadius: '10px',
-  padding: '1rem',
-};
-
-const rowStyle = {
-  display: 'flex',
-  justifyContent: 'space-evenly',
-  alignItems: 'center',
-  borderBottom: '1px solid lightgray',
-};
-
-const minTempStyle = {
-  color: 'blue',
-};
-
-const maxTempStyle = {
-  color: 'red',
-};
+import styles from './FiveDay.module.css';
 
 const FiveDayForecast = (props) => {
   const weatherData = props.weatherData.list;
   const { tempUnit } = props;
+
   // console.log('five: ', weatherData);
+
   let dateOptions = {
     weekday: 'long',
-    // year: 'numeric',
     month: 'long',
     day: 'numeric',
   };
 
   return (
-    <div className='fiveDayList' style={fiveDayStyle}>
+    <div className={styles.FiveDayContainer}>
       <h3>Five day Forecast</h3>
       {weatherData.map((weatherData) => (
-        <div key={weatherData.dt} className='row' style={rowStyle}>
+        <div key={weatherData.dt} className={styles.row}>
           {/* Timestamp into readable date */}
           <span>
             {/* För svenska 'sv-SE' */}
@@ -64,7 +37,7 @@ const FiveDayForecast = (props) => {
           {weatherData.weather.map((weatherType) => (
             <div key={Math.random() * 99999}>
               <img
-                className='dayWeatherIcon'
+                className={styles.dayWeatherIcon}
                 src={`http://openweathermap.org/img/wn/${weatherType.icon}@2x.png`}
                 alt={weatherType.description}
                 title={weatherType.description}
@@ -76,12 +49,12 @@ const FiveDayForecast = (props) => {
 
           {/* Temperature */}
           {/* Min Temperature */}
-          <span className='minTemp' style={minTempStyle}>
+          <span className={styles.minTemp} >
             {'⬇' + Math.round(weatherData.main.temp_min)}
             {tempUnit === 'metric' ? '\u00b0C' : '\u00b0F'}
 
             {/* Max Temperature */}
-            <span className='maxTemp' style={maxTempStyle}>
+            <span className={styles.maxTemp} >
               &nbsp; {'⬆' + Math.round(weatherData.main.temp_max)}
               {tempUnit === 'metric' ? '\u00b0C' : '\u00b0F'}
             </span>
